Update v3
Created a menu_bar or toolbar. A basic toolbar for now, will be working
on making it more pretty and attractive. Also may add or remove features
to enhance user experience.

I also updated my string file, to store the values.

**Please take note that, the toasts being displayed is just a test, this
will not be implemented in the code for obvious reasons.**