Update v4
In this update, i have created an UpButton which when the user clicks,
it will prompt them with a confirmation message asking if they would
like exit the app. I also implemted this on the back key on the phone.

Currently, i am working on a database to store user information for the
login and register screen. After discussing it with the group members,
we felt it was important to have as it would allow the user to build
their profile.