Update v7
In this update, I have interated the online database. Hennok has also
added the functionality of the joystick feature,  all the functionality
is working, except for the Camera. There are a few bugs that we still
need to fix, once that has completed we will be able to upload our Beta
App v1.0. Stay tuned.