Update by Tanav Sharma
I ran into some problems, it would not let me load my branch to make
commits. Therefore I am making the commit to the master.

In this update, i created the Navigation Drawer, and added the
functionality of it. I also made some minor UI adjestments.

I also updated the Bio, which allows the user to upload a description to
a online database. The problem I am running into is that it does not
connect to the database, to save the bio because I have to code the
usage premisions in my database.